:wrench: UPDATE: Make tests optional
Added new CLI option ("--runapi") to pytest to render the tests that evaluate the response of an API call optional.

This is because API responses aren't idempotent and can vary especially from API to API making these tests unstable as well. Moreover, these tests don't contain themselves to the core TransLaTeX program but also test external factors which is extra and not essential.