:bug: BUGFIX: Fix tokenizer bug
Fixed major bug causing multi argument (having "{}" more than once) LaTeX commands to not tokenize and detokenize correctly causing more than only the text to be sent to translation and thus producing an incorrect LaTeX output that doesn't compile in the event of an alteration from the translator.

Also fixed an idempotency issue where additional whitespace would be added between the multiple arguments of a LaTeX command not preserving the original string.